libs/tools/generator/core/src/policies/passphrase-least-privilege.spec.ts
import { PolicyType } from "@bitwarden/common/admin-console/enums";
// FIXME: use index.ts imports once policy abstractions and models
// implement ADR-0002
import { Policy } from "@bitwarden/common/admin-console/models/domain/policy";
import { PolicyId } from "@bitwarden/common/types/guid";

import { DisabledPassphraseGeneratorPolicy } from "../data";

import { passphraseLeastPrivilege } from "./passphrase-least-privilege";

function createPolicy(
  data: any,
  type: PolicyType = PolicyType.PasswordGenerator,
  enabled: boolean = true,
) {
  return new Policy({
    id: "id" as PolicyId,
    organizationId: "organizationId",
    data,
    enabled,
    type,
  });
}

describe("passphraseLeastPrivilege", () => {
  it("should return the accumulator when the policy type does not apply", () => {
    const policy = createPolicy({}, PolicyType.RequireSso);

    const result = passphraseLeastPrivilege(DisabledPassphraseGeneratorPolicy, policy);

    expect(result).toEqual(DisabledPassphraseGeneratorPolicy);
  });

  it("should return the accumulator when the policy is not enabled", () => {
    const policy = createPolicy({}, PolicyType.PasswordGenerator, false);

    const result = passphraseLeastPrivilege(DisabledPassphraseGeneratorPolicy, policy);

    expect(result).toEqual(DisabledPassphraseGeneratorPolicy);
  });

  it.each([
    ["minNumberWords", 10],
    ["capitalize", true],
    ["includeNumber", true],
  ])("should take the %p from the policy", (input, value) => {
    const policy = createPolicy({ [input]: value });

    const result = passphraseLeastPrivilege(DisabledPassphraseGeneratorPolicy, policy);

    expect(result).toEqual({ ...DisabledPassphraseGeneratorPolicy, [input]: value });
  });
});
